+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- Lighter npm distribution. Ignore not needed resources.
+
 ### Changed
 - Change logo.
 - Documentation fixes.